Refactor `run_silence_injector` into `bincommon`
This deduplicates `run_silence_injector` from being copy and pasted
across two files.
